08292025_SANCHRA_GCP-37/38/39/40

+++ capstone/container_management_system/src/main/java/com/g3w4/container_management_system/model/Customer.java
@@ -16,10 +16,11 @@ import lombok.NoArgsConstructor;
 @Entity
 @Table(name = "customer")
 public class Customer {
+
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Column(name = "customer_id")
-    private long customerId;
+    private int customerId;
 
     @Column(name = "first_name")
     private String firstName;
