+++ capstone/container_management_system/src/main/java/com/g3w4/container_management_system/model/Roles.java
@@ -20,7 +20,7 @@ public class Roles {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Column(name = "role_id")
-    private long roleId;
+    private int roleId;
 
     @Column(name = "role_name")
     private String roleName;
