+++ capstone/container_management_system/src/main/java/com/g3w4/container_management_system/model/TrackingStatus.java
@@ -20,7 +20,7 @@ public class TrackingStatus {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Column(name = "status_id")
-    private long statusId;
+    private int statusId;
 
     @Column(name = "status_name")
     private String statusName;
